# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Stixor Office Multi-Agent API",
        "version": "0.2.0"
    },
    "paths": {
...
        "/chat": {
            "post": {
                "tags": [
                    "chat"
                ],
                "summary": "Chat",
                "operationId": "chat_chat_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChatRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/chat/stream": {
            "post": {
                "tags": [
                    "chat"
                ],
                "summary": "Chat Stream",
                "description": "Server-sent events: `route` once the supervisor decides, then `message` with the reply and artifacts.",
                "operationId": "chat_stream_chat_stream_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChatRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/threads": {
            "get": {
                "tags": [
                    "chat"
                ],
                "summary": "My Threads",
                "operationId": "my_threads_threads_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/threads/{thread_id}": {
            "get": {
                "tags": [
                    "chat"
                ],
                "summary": "Thread History",
                "operationId": "thread_history_threads__thread_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "thread_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Thread Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ThreadHistory"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/clients": {
            "get": {
                "tags": [
                    "finance"
                ],
                "summary": "Clients",
                "operationId": "clients_clients_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/invoices": {
            "get": {
                "tags": [
                    "finance"
                ],
                "summary": "Invoices",
                "operationId": "invoices_invoices_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "client",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "",
                            "title": "Client"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/finance/bank-accounts": {
            "get": {
                "tags": [
                    "finance"
                ],
                "summary": "Bank Accounts",
                "operationId": "bank_accounts_finance_bank_accounts_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ChatRequest": {
                "properties": {
                    "message": {
                        "type": "string",
                        "maxLength": 20000,
                        "minLength": 1,
                        "title": "Message"
                    },
                    "thread_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Thread Id"
                    },
                    "department": {
                        "type": "string",
                        "enum": [
                            "auto",
                            "hr",
                            "devops",
                            "finance",
                            "pm",
                            "developer"
                        ],
                        "title": "Department",
                        "default": "auto"
                    },
                    "template_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Template Id"
                    }
                },
                "type": "object",
                "required": [
                    "message"
                ],
                "title": "ChatRequest"
            },
            "ChatResponse": {
                "properties": {
                    "thread_id": {
                        "type": "string",
                        "title": "Thread Id"
                    },
                    "reply": {
                        "type": "string",
                        "title": "Reply"
                    },
                    "agent": {
                        "type": "string",
                        "title": "Agent"
                    },
                    "route_reason": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Route Reason"
                    },
                    "artifacts": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Artifacts",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "thread_id",
                    "reply",
                    "agent"
                ],
                "title": "ChatResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "HistoryMessage": {
                "properties": {
                    "role": {
                        "type": "string",
                        "enum": [
                            "user",
                            "assistant"
                        ],
                        "title": "Role"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content"
                    },
                    "agent": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Agent"
                    },
                    "artifacts": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Artifacts",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "role",
                    "content"
                ],
                "title": "HistoryMessage"
            },
...
            "ThreadHistory": {
                "properties": {
                    "thread_id": {
                        "type": "string",
                        "title": "Thread Id"
                    },
                    "messages": {
                        "items": {
                            "$ref": "#/components/schemas/HistoryMessage"
                        },
                        "type": "array",
                        "title": "Messages"
                    }
                },
                "type": "object",
                "required": [
                    "thread_id",
                    "messages"
                ],
                "title": "ThreadHistory"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}